Setup the initial E2E (minus orchestrator/caching) for the new system of IReportItemProvider, DataSources, and the TaskOrchestrator
Adds the implementation for NuGetRecentlyUpdated check (maybe should be called NuGetRecentlyUpdatedChecker?) to see if package has been updated in last year.
Cleans-up/clarifies some of the NuGetPackageRegistration class names
Rough display of check in UI to see it working
TODO: Need to shuffle code around to proper spots vs. being hard-coded in the MainViewModel so we can do status reporting on thread tasks and caching, etc...