Split out Project Info from the Package, a Package can now have multiple project references
This lets us just have a single high-level view of the packages vs. having them duplicated in the main list.
This also means we won't duplicate reports/checks that don't rely on Framework information (and they can check that all at once against the list of frameworks for projects for that package anyway this way).