Address @erinpwoo's comment about deselecting the package from the list when viewing the log
Adds tooltip to button
Discussed alternatives for icon for button between Book (Dictionary) and VS Output Window type icon, sticking with List.
Logic is a bit complicated due to states and interactions, but tested now. We should reselect the prior item if you just toggle the view, and we should properly close the view and show the new item if you click elsewhere instead.